feat: add player follow camera
