--- Void Runner/Assets/Scripts/CinemachineInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 701d7fbf6bbc07c4785e636f44d95449
